requirements.txt
# Core data science
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
xgboost==2.1.3
lightgbm==4.5.0
shap==0.46.0
imbalanced-learn==0.12.4

# LLM
anthropic==0.40.0

# API
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3

# Frontend
streamlit==1.40.2
plotly==5.24.1

# Database
sqlalchemy==2.0.36

# Config
python-dotenv==1.0.1

# Visualization
matplotlib==3.9.3
seaborn==0.13.2

# Testing
pytest==8.3.4
pytest-cov==6.0.0

# Utilities
requests==2.32.3
